+++ packages/backend/src/nest/storage/orbitDb/orbitDb.service.ts
@@ -10,7 +10,6 @@ import { MessagesAccessController } from '../channels/messages/orbitdb/MessagesA
 import {
   createOrbitDB,
   type OrbitDBType,
-  type IdentitiesType,
   useAccessController as orbitDbUseAccessController, // this is to fix a linting issue about react hooks
   ComposedStorage,
   LRUStorage,
@@ -19,21 +18,23 @@ import {
   LogEntry,
   Entry,
   DatabaseType,
+  LogType,
+  IdentitiesType,
 } from '@orbitdb/core'
 import { HeliaLibp2p } from 'helia'
 import { OrbitDbStorage } from '../../types'
-import { IdentitiesWithStorage } from './identitiesWithStorage'
 import drain from 'it-drain'
 import IPFSBlockStorage from './ipfsBlockStorage'
 import { LocalDbService } from '../../local-db/local-db.service'
 import { SigChainService } from '../../auth/sigchain.service'
 import { LogEntrySyncMessage } from '../../qss/qss.types'
+import { LFAIdentities } from './identity/lfa/lfa-identity.service'
 
 @Injectable()
 export class OrbitDbService {
   private orbitDbInstance: OrbitDBType | undefined = undefined
-  private stores: Record<string, any> = {}
-  public identities: IdentitiesType | undefined = undefined
+  private stores: Record<string, DatabaseType> = {}
+  public identities: LFAIdentities | undefined = undefined
   public static readonly events = new EventEmitter()
 
   private readonly logger = createLogger(OrbitDbService.name)
@@ -41,7 +42,8 @@ export class OrbitDbService {
   constructor(
     @Inject(ORBIT_DB_DIR) public readonly orbitDbDir: string,
     private readonly localDbService: LocalDbService,
-    private readonly sigChainService: SigChainService
+    private readonly sigChainService: SigChainService,
+    private readonly lfaIdentities: LFAIdentities
   ) {
     OrbitDbService.events.on('update', (entry: LogEntry) => {
       if (entry.identity == this.orbitDbInstance?.identity.hash) {
@@ -68,14 +70,17 @@ export class OrbitDbService {
 
     orbitDbUseAccessController(MessagesAccessController)
 
-    this.identities = await IdentitiesWithStorage(this.orbitDbDir, ipfs)
+    /**
+     * This overrides the built-in identity system to use our custom LFA-based identity service
+     */
+    this.identities = this.lfaIdentities
 
     const peerId = ipfs.libp2p.peerId
     const orbitDb = await createOrbitDB({
       ipfs,
       id: peerId.toString(),
       directory: this.orbitDbDir,
-      identities: this.identities,
+      identities: this.identities as any, // our type diverges from the base type
     })
 
     this.orbitDbInstance = orbitDb
@@ -252,7 +257,14 @@ export class OrbitDbService {
       throw new Error('OrbitDB instance is not initialized. Call create() first.')
     }
 
-    return []
+    const entries: LogEntry[] = []
+    const store = this.stores[address]
+    for (const hash of hashes) {
+      const entry = await (store.log as LogType).get(hash)
+      entries.push(entry)
+    }
+
+    return entries
   }
 
   public static async createDefaultStorage(